Major update
I stole the codes mainly from SEOwnedDE and other sources, though the sigs are mine (yes I learned how to sig recently and this is supposed to be a flex). I am not able to dedicate my time to gaming/coding as I'll be focusing on my university studies, so I don't exactly remember the state the cheat is in or what works and what doesn't.